monorepo/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
motor>=3.3.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.0
